chore: rename schema util
