+++ src/oas2/transformers/getExamplesFromSchema.ts
